requirements.txt
# Core dependencies
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
pydantic>=2.0.0

# Hardware interface
pyserial>=3.5

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Security
bandit>=1.7.0
safety>=2.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Agent framework (optional, for advanced features)
# langchain>=0.1.0
# openai>=1.0.0

# Performance & Async (Phase 2)
aiohttp>=3.8.0

# Monitoring & Logging (Phase 2)
psutil>=5.9.0

# Resilience & Error Handling (Phase 2)
# tenacity>=8.2.0  # For advanced retry patterns
